Attempting to test an addition of a config setting
I'm attempting to write a method which will take in an array of
parameters and use them to set the appropriate cosmetic details
when these are explicitly passed in from the command line, or to
simply pass the parameters to the cosmetic details labels in the
same order in which they are entered.  However, the behaviour is
proving more complicated to implement than I first anticipated.