Character cosmetic details (#2)
* Cosmetic Details Terminology added.

This will be used by the display and editing panels in the GUI.

* Display Panel for Cosmetic Details written.

* Editor for Cosmetic Details Written

Also, a big refactor on all classes to bring them inline with Java
coding conventions, particularly those on naming.

* Cosmetic Character Details Running

They now display nicely and are editable.

* Character Name may be set from command line at application launch.

* Attempting to test an addition of a config setting

I'm attempting to write a method which will take in an array of
parameters and use them to set the appropriate cosmetic details
when these are explicitly passed in from the command line, or to
simply pass the parameters to the cosmetic details labels in the
same order in which they are entered.  However, the behaviour is
proving more complicated to implement than I first anticipated.

* Config setting logic moved to Cosmetic Details model.

* setConfigDetails now working correctly

Named details are set, provided the intended value always follows
the intended key.  Unnamed are simply set in the same order as the
labels appear in the character sheet.